AD-64: Privatkunde ändert Daten eines Spots

+++ src/main/java/de/hs/furtwangen/bam/spots/repository/SpotRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import de.hs.furtwangen.bam.spots.model.Spot;
 
-public interface SpotRepository extends CrudRepository<Spot, Long>{
+public interface SpotRepository extends CrudRepository<Spot, Integer>{
 
 	public Spot findByName(String name);
